retrieve and HDF5 representation of an embedding produced by pbg in simba

retrieve and HDF5 representation of an embedding produced by pbg in simba

get_emb_h5(
  ctarpath,
  type = "cell",
  numtag = "_0",
  gname = "graph0",
  mname = "model",
  spec = "v10",
  exdir = tempdir()
)

Arguments

ctarpath

character(1) path to a compressed tar of pbg folder

type

character(1) 'cell' or 'gene'

numtag

character(1) defaults to "_0"

gname

character(1) used as dirname in build_and_train_pbg

mname

character(1) used as output in build_and_train_pbg

spec

character(1) defaults to "v10"

exdir

character(1) path where tar operation outputs will be placed, defaults to tempdir()

Value

result of rhdf5::h5read on the embedding values

Examples

tpath = system.file(file.path("extdata", "pbg3k.tar.xz"), package="Simba4Bioc")
cemb = get_emb_h5(tpath)
dim(cemb)
#> [1]   50 2700